Add support for token_endpoint_auth_method = private_key_jwt
This isn't particularly well documented by Okta. You'll see it listed
under https://developer.okta.com/docs/reference/api/apps/#credentials -
and an example including the JWKS elements is available here:
https://developer.okta.com/docs/reference/api/apps/#request-example-10

The general idea behind this is described at
https://developer.okta.com/docs/reference/api/oidc/#jwt-with-private-key
- basically, a private key is held client side. The associated public
key is provided to Okta. Client authentication is achieved by signing a
JWT containing the appropriate client ID. Okta then check the JWT was
signed using one of the keys associated with that client.

As seems to be the case with quite a few things, this isn't covered by
their SDK. I've tried to work around this by defining structs under the
SDK package and embedding the original structs. If this doesn't seem
overly suitable, we'll just have to wait for these bits to be reflected
in the official SDK.

In terms of naming - the JWKS spec uses short key names for various bits
- these do map to suitable english equivalents, but I sort of get the
impression people are mostly used to the short forms.. and requiring
people to map them might be a bit of a nuisance. Happy to change that
if anyone feels strongly.

Some properties are dependent on the key type. For example, with RSA
"e" + "n" are required by the Okta API. Happy to get any pointers on
how to require these in schema _only if the key type is RSA_. I believe
Okta support kty = EC (Elliptic Curve) with "x" + "y" properties - if
so this can be expanded in the future.

Final note - I've used the list type here - it seems the set type would
work too, I'm not sure what the advantages of using one over the other
are in a terraform setting - happy to adjust.