FIX: custom_client_id changes should not recreate resource
This is a change in existing provider behavior. However, its
the only way to ensure that users can migrate from custom_client_id
to client_id without recreating their apps. The other way would
be to edit in statefile, or to remove from state and import, both
of which involve more effort.

I believe this is a reasonable tradeoff considering that the
next release will remove the other attribute.