Add resource definiton for Okta Event Hooks
This adds a new okta_event_hook resource for managing Event Hooks.  The
resource itself is mostly a copy/paste job from the existing
okta_inline_hook resource.  Since the two are so similar I re-worked
sdk/inline_hooks.go to generalize it to work with both event hooks and
inline hooks and renamed it to sdk/hooks.go.