Add resource definiton for Okta Event Hooks (#14)
* Add resource definiton for Okta Event Hooks

This adds a new okta_event_hook resource for managing Event Hooks.  The
resource itself is mostly a copy/paste job from the existing
okta_inline_hook resource.  Since the two are so similar I re-worked
sdk/inline_hooks.go to generalize it to work with both event hooks and
inline hooks and renamed it to sdk/hooks.go.

* Revert CHANGELOG.md

The changelog entries get created by maintainers as part of the
release process.

* Get rid of overly-clever JSON formatting

* Split hooks.go back apart into inline_hooks.go and event_hooks.go

* Clean up and add comments to help explain testCheckResourceSetAttr