Update "id" and "name" attributes of okta_group data source to be computed.
This allows the values to be passed in to these attributes and remain option.
However, Terraform will know the attributes are non-optional attributes when using attributes in other resources.

This avoids errors "The argument "<attribute>" is required, but no definition was found." when using the id attribute of the data source as the value to a required attribute of another resource.

Issue #1642