Handle multiple apps returned by SAML app data source, where one app exactly matches.
In the case where there are two apps: "MyApp" and "MyApp 2".
If the data source for saml apps queried with label "MyApp", if the API returns "MyApp 2" as the first app in the list, the data source will fail.
This change now iterate through all of the apps returned and returning an exact match, if "label" has been specified.